+++ CHANGELOG.md
@@ -4,11 +4,12 @@
 ---
 
 - Made contact circles themeable
-- Added Adwaita Light/Dark themes
 - Added Catppuccin Mocha theme
 - Added Catppuccin Latte theme
 - Added Dracula theme
 - Added Toky Night theme
+- Added Adwaita Light/Dark themes
+- Added Breeze Light/Dark themes
 
 
 **v0.2.1 - 05-07-2026**